Move read model conversion classes to a better namespace
Since these concepts will only be used by the packages that connect
chimera to any service bus implementation, in middleware that handle the
queries, it makes more sense to keep them under `ServiceBus`.
