+++ src/ServiceBus/ReadModelConverter/Query.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Lcobucci\Chimera\ReadModelConverter;
+namespace Lcobucci\Chimera\ServiceBus\ReadModelConverter;
 
 /**
  * Defines the public API of a query object that uses a callback to convert query results
@@ -10,7 +10,6 @@ interface Query
 {
     /**
      * Retrieves the callback function to be used to convert domain objects into read models
-     *
      */
     public function conversionCallback(): callable;
 }
